# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  handler.genPassResponse:
    properties:
      password:
        type: string
    type: object
  handler.shortenURLRequest:
    properties:
      exp_time:
        minimum: 1
        type: integer
      url:
        type: string
    required:
    - exp_time
    - url
    type: object
  service.HealthStatus:
    properties:
      instance_id:
        description: InstanceId is the unique identifier of this running instance.
        type: string
      message:
        description: Message indicates the overall health state (e.g., "OK").
        type: string
      service_name:
        description: ServiceName is the name of the service as configured.
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple bookmark management API.
  title: Bookmark Management API
  version: "1.0"
paths:
  /gen-pass:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.genPassResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Generate Password
      tags:
      - password
  /health-check:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.HealthStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Health Check
      tags:
      - health-check
  /links/redirect/{code}:
    get:
      parameters:
      - description: Shortened URL code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "301":
          description: Redirect to original URL
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Original URL
      tags:
      - URL Shortener
  /links/shorten:
    post:
      consumes:
      - application/json
      parameters:
      - description: URL to shorten
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/handler.shortenURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.shortenURLRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Shorten URL
      tags:
      - URL Shortener
schemes:
- http
swagger: "2.0"